Rename slice_utils -> utils package
