+++ common/utils/time.go
@@ -1,4 +1,4 @@
-package slice_utils
+package utils
 
 func HoursToUnixSeconds(hours int) int64 {
 	return int64(hours) * 60 * 60
