+++ services/decision/controller/controller.go
@@ -87,7 +87,7 @@ func UpdateDecision(w http.ResponseWriter, r *http.Request) {
 
 	decision.Reviewer = r.Header.Get("HackIllinois-Identity")
 	decision.Timestamp = time.Now().Unix()
-	decision.ExpiresAt = decision.Timestamp + slice_utils.HoursToUnixSeconds(config.DECISION_EXPIRATION_HOURS)
+	decision.ExpiresAt = decision.Timestamp + utils.HoursToUnixSeconds(config.DECISION_EXPIRATION_HOURS)
 	// Finalized is always false, unless explicitly set to true via the appropriate endpoint.
 	decision.Finalized = false
 
@@ -135,7 +135,7 @@ func FinalizeDecision(w http.ResponseWriter, r *http.Request) {
 	latest_decision.Wave = existing_decision_history.Wave
 	latest_decision.Reviewer = r.Header.Get("HackIllinois-Identity")
 	latest_decision.Timestamp = time.Now().Unix()
-	latest_decision.ExpiresAt = latest_decision.Timestamp + slice_utils.HoursToUnixSeconds(config.DECISION_EXPIRATION_HOURS)
+	latest_decision.ExpiresAt = latest_decision.Timestamp + utils.HoursToUnixSeconds(config.DECISION_EXPIRATION_HOURS)
 
 	err = service.UpdateDecision(id, latest_decision)
